window: Move new-tab button into headerbar
Opening a new tab is a very frequent action, which is currently
exposed in the tab bar itself. However that means that it only
becomes available when more than one tab is used. Moving the
control to the headerbar instead makes the action always readily
available and leaves a bit more space for tabs.

https://bugzilla.gnome.org/show_bug.cgi?id=756798