headerbar: Allow opening new windows by control-clicking new-tab button
New-terminal items open in a new tab or window, depending on the corre-
sponding preference. Users can temporarily toggle that by pressing CTRL
when activating the action. Extend that functionality to the new-tab
button in the headerbar as suggested by Egmont, so users can quickly
open new windows as well.

https://bugzilla.gnome.org/show_bug.cgi?id=756798